# added api error handling

+++ Maid.Core/API/BaseController.cs
@@ -32,13 +32,16 @@
 		public async Task<ActionResult<T>> GetItem(Guid id) {
 			var item = await EntityRepository.GetAsync(id);
 			if (item == null) {
-				return NotFound();
+				return NotFound(id);
 			}
 			return item;
 		}
 
 		[HttpPost()]
 		public ActionResult AddItem(T item) {
+			if (item == null) {
+				return BadRequest("Parameter object is null");
+			}
 			EntityRepository.Create(item);
 			EntityRepository.Save();
 			return Ok();
@@ -46,6 +49,9 @@
 
 		[HttpPut()]
 		public ActionResult EditItem(T item) {
+			if (item == null) {
+				return BadRequest("Parameter object is null");
+			}
 			EntityRepository.Update(item);
 			EntityRepository.Save();
 			return Ok();
@@ -53,6 +59,9 @@
 
 		[HttpDelete()]
 		public ActionResult DeleteItem(T item) {
+			if (item == null) {
+				return BadRequest("Parameter object is null");
+			}
 			EntityRepository.Delete(item);
 			EntityRepository.Save();
 			return Ok();
@@ -60,9 +69,13 @@
 
 		[HttpDelete("{id}")]
 		public ActionResult DeleteItem(Guid id) {
-			EntityRepository.Delete(id);
-			EntityRepository.Save();
-			return Ok();
+			try {
+				EntityRepository.Delete(id);
+				EntityRepository.Save();
+				return Ok();
+			} catch (KeyNotFoundException) {
+				return NotFound(id);
+			}
 		}
 	}
 }
